web/requirements.txt
Django==4.0.3
django-ckeditor==6.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
